refactoring: move Layout from Pages folder
